refactor: Merge solana-runtime and solana-runtime-api packages
