+++ frame/solana/runtime-api/Cargo.toml
@@ -1,21 +1,44 @@
 [package]
 name = "solana-runtime-api"
-version = "0.4.0"
-authors = ["Haderech Pte. Ltd."]
-edition = "2021"
 license = "Apache-2.0"
-repository = "https://github.com/noirhq/noir.git"
+authors = { workspace = true }
+version = { workspace = true }
+edition = { workspace = true }
+repository = { workspace = true }
 publish = false
 
 [dependencies]
-sp-api = { git = "https://github.com/paritytech/polkadot-sdk", branch = "stable2409", default-features = false }
-parity-scale-codec = { workspace = true, default-features = false, features = [ "derive" ] }
-scale-info = { workspace = true, default-features = false, features = [ "derive" ] }
+frame-support = { workspace = true, default-features = false }
+nostd = { workspace = true, features = ["alloc"] }
+pallet-solana = { workspace = true, default-features = false }
+parity-scale-codec = { workspace = true, default-features = false, features = [
+    "derive",
+] }
+scale-info = { workspace = true, default-features = false, features = [
+    "derive",
+] }
+serde = { workspace = true, default-features = false }
+serde_json = { workspace = true, default-features = false }
+solana-account-decoder = { workspace = true, default-features = false }
+solana-compute-budget = { workspace = true, default-features = false }
+solana-inline-spl = { workspace = true, default-features = false }
+solana-rpc-client-api = { workspace = true, default-features = false }
+solana-sdk = { workspace = true, default-features = false }
+sp-api = { workspace = true, default-features = false }
 
 [features]
 default = ["std"]
 std = [
-    "sp-api/std", 
+    "frame-support/std",
+    "pallet-solana/std",
     "parity-scale-codec/std",
     "scale-info/std",
+    "serde/std",
+    "serde_json/std",
+    "solana-account-decoder/std",
+    "solana-compute-budget/std",
+    "solana-inline-spl/std",
+    "solana-rpc-client-api/std",
+    "solana-sdk/std",
+    "sp-api/std",
 ]
