--- frame/solana/runtime/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "solana-runtime"
-version = "0.4.0"
-authors = ["Haderech Pte. Ltd."]
-edition = "2021"
-license = "Apache-2.0"
-repository = "https://github.com/noirhq/noir.git"
-publish = false
-
-[dependencies]
-frame-support = { workspace = true, default-features = false }
-nostd = { workspace = true, default-features = false, features = ["alloc"] }
-pallet-solana = { workspace = true, default-features = false }
-serde = { workspace = true, default-features = false }
-serde_json = { version = "1.0", default-features = false }
-solana-account-decoder = { workspace = true, default-features = false }
-solana-compute-budget = { workspace = true, default-features = false }
-solana-inline-spl = { workspace = true, default-features = false }
-solana-rpc-client-api = { workspace = true, default-features = false }
-solana-runtime-api = { workspace = true, default-features = false }
-solana-sdk = { workspace = true, default-features = false }
-
-[features]
-default = ["std"]
-std = [
-    "frame-support/std",
-    "pallet-solana/std",
-    "serde/std",
-    "serde_json/std",
-    "solana-account-decoder/std",
-    "solana-compute-budget/std",
-    "solana-inline-spl/std",
-    "solana-rpc-client-api/std",
-    "solana-runtime-api/std",
-    "solana-sdk/std",
-]
